requirements.txt
telethon==1.40.0
python_dotenv==1.1.1
fastapi==0.116.1
uvicorn[standard]==0.35.0
motor==3.7.1
psutil==7.1.0